modules/eks/ray-operator/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
attrs==23.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.211
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.3
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.1.0
    # via aws-cdk-lib
aws-cdk-cloud-assembly-schema==38.0.1
    # via aws-cdk-lib
aws-cdk-lambda-layer-kubectl-v29==2.0.0
    # via -r requirements.in
aws-cdk-lib==2.166.0
    # via
    #   -r requirements.in
    #   aws-cdk-lambda-layer-kubectl-v29
cattrs==22.1.0
    # via jsii
constructs==10.0.91
    # via
    #   aws-cdk-lambda-layer-kubectl-v29
    #   aws-cdk-lib
importlib-resources==5.12.0
    # via jsii
jsii==1.105.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lambda-layer-kubectl-v29
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lambda-layer-kubectl-v29
    #   aws-cdk-lib
    #   constructs
    #   jsii
pydantic==2.7.3
    # via
    #   -r requirements.in
    #   pydantic-settings
pydantic-core==2.18.4
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via jsii
python-dotenv==0.21.1
    # via pydantic-settings
pyyaml==6.0.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lambda-layer-kubectl-v29
    #   aws-cdk-lib
    #   jsii
typing-extensions==4.6.3
    # via
    #   jsii
    #   pydantic
    #   pydantic-core
